pyproject.toml
[project]
name = "nl-query-agent"
version = "0.1.0"
description = "Two-agent LangGraph system for natural language querying over the DVD Rental PostgreSQL database"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi>=0.104.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "uvicorn[standard]>=0.24.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    "langchain-mcp-adapters>=0.1.0",
    "mcp>=1.0.0",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.38.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
